moved new permit precompile tests to their own file
