+++ substrate/frame/assets/precompiles/src/lib.rs
@@ -53,6 +53,8 @@ mod migration_tests;
 #[cfg(test)]
 mod mock;
 #[cfg(test)]
+mod permit_precompile_tests;
+#[cfg(test)]
 mod permit_tests;
 #[cfg(test)]
 mod test_helpers;
